fix(prefer-tacit): handling of member expressions and be more strict with optional parameters
fix #486
